Create IcebergTestUtils, test source already shared with extensions-iceberg-s3
