+++ extensions/iceberg/s3/src/test/java/io/deephaven/iceberg/util/S3WarehouseSqliteCatalogBase.java
@@ -6,7 +6,7 @@ package io.deephaven.iceberg.util;
 import io.deephaven.engine.table.Table;
 import io.deephaven.engine.util.TableTools;
 import io.deephaven.extensions.s3.S3Instructions;
-import io.deephaven.iceberg.base.IcebergUtils;
+import io.deephaven.iceberg.base.IcebergTestUtils;
 import io.deephaven.iceberg.junit5.SqliteCatalogBase;
 import io.deephaven.iceberg.sqlite.SqliteHelper;
 import org.apache.iceberg.CatalogProperties;
@@ -133,7 +133,7 @@ abstract class S3WarehouseSqliteCatalogBase extends SqliteCatalogBase {
                 .build());
 
         // Verify all data files have the right scheme
-        final List<DataFile> dataFiles = IcebergUtils.allDataFiles(icebergTable, icebergTable.currentSnapshot())
+        final List<DataFile> dataFiles = IcebergTestUtils.allDataFiles(icebergTable, icebergTable.currentSnapshot())
                 .collect(Collectors.toList());
         assertThat(dataFiles).hasSize(2);
         assertThat(dataFiles).allMatch(dataFile -> dataFileUri(icebergTable, dataFile).getScheme().equals(scheme));
@@ -159,7 +159,7 @@ abstract class S3WarehouseSqliteCatalogBase extends SqliteCatalogBase {
         icebergTable.newAppend().appendFile(newDataFile).commit();
 
         // Verify the new data file has the right scheme
-        final List<DataFile> newDataFiles = IcebergUtils.allDataFiles(icebergTable, icebergTable.currentSnapshot())
+        final List<DataFile> newDataFiles = IcebergTestUtils.allDataFiles(icebergTable, icebergTable.currentSnapshot())
                 .collect(Collectors.toList());
         int s3DataFiles = 0;
         int nonS3DataFiles = 0;
