+++ extensions/iceberg/src/main/java/io/deephaven/iceberg/base/IcebergUtils.java
@@ -9,13 +9,10 @@ import io.deephaven.engine.table.TableDefinition;
 import io.deephaven.engine.table.impl.locations.TableDataException;
 import io.deephaven.iceberg.relative.RelativeFileIO;
 import io.deephaven.iceberg.util.IcebergReadInstructions;
-import io.deephaven.util.annotations.TestUseOnly;
 import org.apache.iceberg.DataFile;
-import org.apache.iceberg.ManifestFiles;
 import org.apache.iceberg.PartitionField;
 import org.apache.iceberg.PartitionSpec;
 import org.apache.iceberg.Schema;
-import org.apache.iceberg.Snapshot;
 import org.apache.iceberg.Table;
 import org.apache.iceberg.catalog.Catalog;
 import org.apache.iceberg.catalog.Namespace;
@@ -28,8 +25,6 @@ import org.apache.iceberg.types.Types;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.io.IOException;
-import java.io.UncheckedIOException;
 import java.math.BigDecimal;
 import java.net.URI;
 import java.time.Instant;
@@ -42,8 +37,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
-import java.util.stream.StreamSupport;
 
 public final class IcebergUtils {
 
@@ -62,36 +55,6 @@ public final class IcebergUtils {
         DH_TO_ICEBERG_TYPE_MAP.put(LocalTime.class, Types.TimeType.get());
         DH_TO_ICEBERG_TYPE_MAP.put(byte[].class, Types.BinaryType.get());
         // TODO (deephaven-core#6327) Add support for more types like ZonedDateTime, Big Decimals, and Lists
-    }
-
-    /**
-     * Get a stream of all {@link DataFile} objects from the given {@link Table} and {@link Snapshot}.
-     *
-     * @param table The {@link Table} to retrieve data files for.
-     * @param snapshot The {@link Snapshot} to retrieve data files from.
-     *
-     * @return A stream of {@link DataFile} objects.
-     */
-    @TestUseOnly
-    public static Stream<DataFile> allDataFiles(@NotNull final Table table, @NotNull final Snapshot snapshot) {
-        return snapshot.allManifests(table.io())
-                .stream()
-                .map(x -> ManifestFiles.read(x, table.io()))
-                .flatMap(IcebergUtils::toStream);
-    }
-
-    /**
-     * Convert a {@link org.apache.iceberg.io.CloseableIterable} to a {@link Stream} that will close the iterable when
-     * the stream is closed.
-     */
-    private static <T> Stream<T> toStream(final org.apache.iceberg.io.CloseableIterable<T> iterable) {
-        return StreamSupport.stream(iterable.spliterator(), false).onClose(() -> {
-            try {
-                iterable.close();
-            } catch (final IOException e) {
-                throw new UncheckedIOException(e);
-            }
-        });
     }
 
     private static String path(@NotNull final String path, @NotNull final FileIO io) {
